requirements.txt
pytorch
pandas
scikit-learn
pickle
numpy
matplotlib
seaborn
tensorflow==2.15